create card store and split user store
