+++ src/assets/main.css
@@ -5,7 +5,7 @@
 body {
   background-color: #0a284e;
   color: #d5d5d5;
-  margin: 0rem 1rem;
+  margin: 0rem 4rem;
 }
 
 a {
